Revert "Compress NodeData and Data"
This reverts commit d969ab22b9bf23c91a3e8d2f39c42dae4d61bef2.
