+++ modules/datastore/src/EventSubscriber/Subscriber.php
@@ -2,12 +2,13 @@
 
 namespace Drupal\datastore\EventSubscriber;
 
-use Drupal\common\Events\Event;
 use Drupal\common\Resource;
 use Drupal\common\LoggerTrait;
+use Drupal\metastore\Events\DatasetUpdate;
 use Drupal\metastore\Events\Registration;
 use Drupal\metastore\LifeCycle\LifeCycle;
 use Drupal\metastore\ResourceMapper;
+use Drupal\metastore\Storage\MetastoreEntityStorageInterface;
 use Symfony\Component\EventDispatcher\EventSubscriberInterface;
 
 /**
@@ -64,11 +65,11 @@ class Subscriber implements EventSubscriberInterface {
   /**
    * Purge resources.
    *
-   * @param \Drupal\common\Events\Event $event
+   * @param \Drupal\metastore\Events\DatasetUpdate $event
    *   Dataset publication.
    */
-  public function purgeResources(Event $event) {
-    $item = $event->getData();
+  public function purgeResources(DatasetUpdate $event) {
+    $item = $event->getItem();
 
     /** @var \Drupal\datastore\Service\ResourcePurger $resourcePurger */
     $resourcePurger = \Drupal::service('dkan.datastore.service.resource_purger');
