+++ modules/metastore/modules/metastore_search/tests/src/Unit/Plugin/search_api/datasource/DkanDatasetTest.php
@@ -6,7 +6,7 @@ use Drupal\Core\DependencyInjection\Container;
 use Drupal\Core\Entity\EntityStorageInterface;
 use Drupal\Core\Entity\EntityTypeManager;
 use Drupal\Core\Entity\Query\QueryInterface;
-use Drupal\metastore\Storage\NodeData;
+use Drupal\metastore\Storage\Data;
 use Drupal\metastore\Storage\DataFactory;
 use Drupal\metastore_search\Plugin\search_api\datasource\DkanDataset;
 use Drupal\node\NodeInterface;
@@ -48,8 +48,8 @@ class DkanDatasetTest extends TestCase {
       ->add(QueryInterface::class, 'execute', $executeSequence)
       ->add(QueryInterface::class, 'range', QueryInterface::class)
       ->add(EntityTypeRepository::class, 'getEntityTypeFromClass', NULL)
-      ->add(DataFactory::class, 'getInstance', NodeData::class)
-      ->add(NodeData::class, 'retrievePublished', '{}')
+      ->add(DataFactory::class, 'getInstance', Data::class)
+      ->add(Data::class, 'retrievePublished', '{}')
       ->getMock();
 
     \Drupal::setContainer($container);
