+++ modules/metastore/tests/src/Storage/DataTest.php
@@ -3,7 +3,7 @@
 namespace Drupal\Tests\metastore\Storage;
 
 use Drupal\Core\Entity\EntityTypeManager;
-use Drupal\metastore\Storage\NodeData;
+use Drupal\metastore\Storage\Data;
 use Drupal\node\NodeStorage;
 use MockChain\Chain;
 use PHPUnit\Framework\TestCase;
@@ -21,7 +21,7 @@ class DataTest extends TestCase {
       ->add(EntityTypeManager::class, 'getStorage', NodeStorage::class)
       ->getMock();
 
-    $data = new NodeData('dataset', $etm);
+    $data = new Data('dataset', $etm);
     $this->assertInstanceOf(NodeStorage::class, $data->getEntityStorage());
   }
