+++ modules/metastore/tests/src/WebServiceApiTest.php
@@ -5,7 +5,7 @@ namespace Drupal\Tests\metastore;
 use Drupal\metastore\Exception\ExistingObjectException;
 use Drupal\metastore\Exception\MissingObjectException;
 use Drupal\metastore\Exception\UnmodifiedObjectException;
-use Drupal\metastore\Storage\NodeData;
+use Drupal\metastore\Storage\Data;
 use Drupal\metastore\Service;
 use Drupal\metastore\WebServiceApi;
 use Drupal\metastore\SchemaRetriever;
@@ -261,8 +261,8 @@ EOF;
    */
   public function testPatchBadPayload() {
     $mockChain = $this->getCommonMockChain();
-    $mockChain->add(NodeData::class, 'retrieve', "{ }");
-    $mockChain->add(NodeData::class, 'store', new \Exception("Could not store"));
+    $mockChain->add(Data::class, 'retrieve', "{ }");
+    $mockChain->add(Data::class, 'store', new \Exception("Could not store"));
     $mockChain->add(RequestStack::class, 'getCurrentRequest', Request::class);
     $mockChain->add(Request::class, 'getContent', "{");
     $mockChain->add(Request::class, 'getRequestUri', "http://blah");
